permission [nfc]: Move some permissions code out to new file
